% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fifty_old.r
\name{calc50crossing}
\alias{calc50crossing}
\title{Calculate the time of 50\% abundance of centroids of the clusters.}
\usage{
calc50crossing(clustered)
}
\arguments{
\item{clustered}{A fclust object containing the clustering details.}
}
\value{
A matrix, where each row represents an event. It contains for each event, the cluster number, (x (time), y (abundance)) coordinates and direction of crossing.
}
\description{
Calculate the time of 50\% abundance of centroids of the clusters.
}
\seealso{
\code{\link[e1071]{cmeans}} for clustering time profiles.
}
