% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterChng.r
\name{calcClusterChng}
\alias{calcClusterChng}
\title{Calculate changes at all intervals within each cluster}
\usage{
calcClusterChng(Tc, clusters)
}
\arguments{
\item{Tc}{A matrix containing time course data.}

\item{clusters}{A vector with same length as number of profiles. The cluster labels are assumed to be numbers, starting from 1.}
}
\value{
A list (with same length as number of clusters) containing the Tukey summaries.
}
\description{
Create and run generalised linear models and post-hoc tukey contrasts at all time points for each cluster.
}
\seealso{
\code{\link[e1071]{cmeans}} for clustering.
}
