% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderedEvents.r
\name{calculateOrder}
\alias{calculateOrder}
\title{Calculate the ordering of events}
\usage{
calculateOrder(Tc, clustered, mat_events, test = "wilcox",
  fdrSignif = 0.05, phosEventTh = 0.5, dephosEventTh = 0.5)
}
\arguments{
\item{Tc}{A matrix containing time course data.}

\item{clustered}{A \code{fclust} object containing the clustering details.}

\item{mat_events}{A matrix containing the time regions (and events for the centroid) generated by running the \code{calcEvents} function.}

\item{test}{The test used to compare the event time distributions. The choice is between a parametric or a non-paramtric test:
\describe{
        \item{\code{wilcox}}{Choosing this will result in event times being compared using the Mann Whitney U test}
        \item{\code{t-test}}{Choosing this will result in event times being compared using the Students t-test}
}}

\item{fdrSignif}{The significant cutoff for FDR correction, performed after pairwise comparisons of the event times.}

\item{phosEventTh}{Value between 0 and 1, to define the threshold at which a phosphorylation (or increasing) event occurs. Here, a threshold of 0 corresponds to the minimal value within an interval. An event identified in this way is designated as 1.}

\item{dephosEventTh}{Value between 0 and 1, to define the threshold at which a dephosphorylation (or decreasing) event occurs. A threshold of 0 corresponds to the maximal value within an interval. An event identified in this way is designated as -1.}
}
\value{
A list with four named objects is returned, which encodes the order of the events and the clusters.
}
\description{
Here, the events identified from the centroids are used to generate time-distributions using profiles in the clusters. From these distributions, a matrix (either by running t-test or wilcox) is generated of the p-values and the t-statistic (or the 95\% confidence interval, respectively). The p-values are FDR corrected and are used to define a directed graph, where events with significant time differences are ordered according to the t-statistic (for t-test) or the 95\% confidence interval (for wilcox test). This matrix is transitively reduced and defines the order.
}
\seealso{
\code{\link[e1071]{cmeans}} for clustering time profiles, \code{\link{calcEvents}} for identifying events.
}
