% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderedEvents_combined.r
\name{calculateOrder_combined}
\alias{calculateOrder_combined}
\title{Calculate the ordering of events}
\usage{
calculateOrder_combined(
  list_Tc,
  list_clusters,
  list_events,
  list_timeMap,
  test = "wilcox",
  fdrSignif = 0.05,
  incrEventTh = 0.5,
  decrEventTh = 0.5
)
}
\arguments{
\item{list_Tc}{List of matrices, where each matrix contains time course data for each 'ome'.}

\item{list_clusters}{List of vectors, one for each 'ome', where the order of the 'ome' should be similar to that in list_Tc. Each of these vectors contains a cluster label (assumed to be numeric and starting from 1). The length and order of the genes in these clusters should also be similar to the order in the matricies provided in list_Tc.}

\item{list_events}{List of matricies, where each matrix contains the time regions (and events for the centroid) generated by running the \code{calcEvents} function.}

\item{list_timeMap}{List of vectors, one for each 'ome', where each vector indicates the time point at which each of the 'omic' data sets were measured. For example, if a phosphoproteomics data set was measured at basal, 2minutes and 5minutes, and a proteomics data set was measured at basal, 3minutes and 5 minutes, then list_timeMap=[c(0,1,3), c(0,2,3)] (phosphoproteomics, then proteomics).}

\item{test}{The test used to compare the event time distributions. The choice is between a parametric or a non-parametric test:
\describe{
        \item{\code{wilcox}}{Choosing this will result in event times being compared using the Mann Whitney U test}
        \item{\code{t-test}}{Choosing this will result in event times being compared using the Students t-test}
}}

\item{fdrSignif}{The significance cutoff for FDR correction, performed after pairwise comparisons of the event times.}

\item{incrEventTh}{Value between 0 and 1, to define the threshold at which a phosphorylation (or increasing) event occurs. Here, a threshold of 0 corresponds to the minimum value within an interval. (Such an event is designated as 1).}

\item{decrEventTh}{Value between 0 and 1, to define the threshold at which a dephosphorylation (or decreasing) event occurs. A threshold of 0 corresponds to the maximum value within an interval. (Such an event is designated as -1).}
}
\value{
A list with four named objects ($mat_events_withOrder, $individEventOrder, $signifs, $test) is returned, which encodes the order of the events and the clusters.
}
\description{
For each event, we generate distributions of time of event occurance using profiles within clusters. These distributions are then pair wise compared and the results of all comparisons are represented in the graph space, where events are the graph nodes, and the directional edges represent that the source node event occurs significantly early than the target node. This graph is then utilized to generate the ordering.
}
\seealso{
\code{\link[e1071]{cmeans}} for clustering time profiles, \code{\link{calcEvents}} for identifying events.
}
