% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterChng.r, R/calcClusterChng_old.r,
%   R/calcClusterChng_v2.r
\name{convertMatToDfForGlmFormat}
\alias{convertMatToDfForGlmFormat}
\title{Convert a list of matrices of each cluster, to a list of data frames (in the required glm format) for each cluster. Timepoint, and profileNum are factors, and experimentalObs (ratio) is y.}
\usage{
convertMatToDfForGlmFormat(list_matrices)

convertMatToDfForGlmFormat(list_matrices)

convertMatToDfForGlmFormat(list_matrices)
}
\arguments{
\item{list_matrices}{A list of matrices for each cluster}

\item{list_matrices}{A list of matrices for each cluster}

\item{list_matrices}{A list of matrices for each cluster}
}
\value{
A list containing glm's for each cluster.

A list containing glm's for each cluster.

A list containing glm's for each cluster.
}
\description{
Convert a list of matrices of each cluster, to a list of data frames (in the required glm format) for each cluster. Timepoint, and profileNum are factors, and experimentalObs (ratio) is y.

Convert a list of matrices of each cluster, to a list of data frames (in the required glm format) for each cluster. Timepoint, and profileNum are factors, and experimentalObs (ratio) is y.

Convert a list of matrices of each cluster, to a list of data frames (in the required glm format) for each cluster. Timepoint, and profileNum are factors, and experimentalObs (ratio) is y.
}
