% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge.r
\docType{data}
\name{ge.temporal.changed}
\alias{ge.temporal.changed}
\title{Microarray gene expression profiles}
\format{An object of class \code{matrix} with 2566 rows and 7 columns.}
\usage{
ge.temporal.changed
}
\description{
This dataset was taken from Ma \emph{et al.} 2014. It contains 2566 profiles (non-duplicated) at 7 time points (including basal).

\itemize{
\item The first column contains the protein name.
\item The second column contains "0" for subsequent standardisation.
\item The remaining columns contain gene-expression ratios.
}
}
\references{
Ma \emph{et al.} 2014. Available at DOI: \url{10.1128/MCB.00583-14}
}
\keyword{data}
