% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{getTimeRegionsWithMaximalChange}
\alias{getTimeRegionsWithMaximalChange}
\title{Calculate time regions}
\usage{
getTimeRegionsWithMaximalChange(
  glmTukeyForEachClus,
  numTps,
  signifTh = 0.05,
  phosZscoreTh = 0,
  dephosZscoreTh = 0
)
}
\arguments{
\item{glmTukeyForEachClus}{A list (with same length as number of clusters) containing the GLM post-hoc Tukey contrasts for all time points. Obtained by running the \code{calcClusterChng} function.}

\item{numTps}{Total number of time points in the data.}

\item{signifTh}{P-value cutoff of the Tukey contrasts.}

\item{phosZscoreTh}{A z-score value filter to remove phosphorylation (or increasing) events, such that all events with a z-score <= phosZscoreTh are removed.}

\item{dephosZscoreTh}{A z-score value filter to remove dephosphorylation (or decreasing) events, such that all events with a z-score >= dephosZscoreTh are removed.}
}
\value{
A List of matrices, one for each cluster, containing the time regions information. The columns in each matrix are cluster, zscore, start time point, end time point, and direction.
}
\description{
Utilize the Tukey significantly constrasting intervals, and determine time regions. Time regions are intervals such that:
1. the z-score is maximal
2. if a region consists of any sub-intervals, they too have the same sign
3. intervals are non-overlapping.
}
\seealso{
\code{\link{calcClusterChng}}, \code{\link{missingStats}}
}
