% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humphrey_noDup.r
\docType{data}
\name{humphrey_noDup}
\alias{humphrey_noDup}
\alias{humphrey.noDup}
\title{Phosphorylation abundance profiles}
\format{An object of class \code{matrix} with 3172 rows and 9 columns.}
\usage{
humphrey.noDup
}
\description{
This dataset was taken from Humphrey \emph{et al.} 2013. It contains 3172 profiles (non-duplicated) at 9 time points (including basal). 

\itemize{
\item The first column contains the protein name and the phospho-site position in the format "proteinName;position".
\item The second column contains "0" for subsequent standardisation.
\item The remaining columns contain abundance ratios.
}
}
\references{
Humphrey \emph{et al.} 2013. Available at: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3690479/}
}
\keyword{data}
