% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{missingStats}
\alias{missingStats}
\title{Summary of clusters providing percentage of profiles with missing events}
\usage{
missingStats(Tc, clustered, mat_fiftyPoints, phosTh, dephosTh)
}
\arguments{
\item{Tc}{A matrix containing time course data.}

\item{clustered}{A fclust object containing the clustering details.}

\item{mat_events}{A matrix containing the event information generated by running the \code{calcEvents} function.}

\item{phosEventTh}{Value between 0 and 1, to define the threshold at which a phosphorylation (or increasing) event occurs. Here, a threshold of 0 corresponds to the minimal value within an interval. An event identified in this way is designated as 1.}

\item{dephosEventTh}{Value between 0 and 1, to define the threshold at which a dephosphorylation (or decreasing) event occurs. A threshold of 0 corresponds to the maximal value within an interval. An event identified in this way is designated as -1.}
}
\value{
Displays a plot and returns a matrix, with z-scores masked (NA) according to the signficance threshold specified.
}
\description{
Summary of clusters: providing information regarding the percentage of profiles which do not contain events within the identified \code{timeRegions}.
}
\seealso{
\code{\link{plotZP}}, \code{\link{summaryGetZP}}, \code{\link{calc50crossing}}
}
