% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.r
\docType{data}
\name{multiomics}
\alias{multiomics}
\alias{phosTsData.stand}
\alias{protTsData.stand}
\alias{rnaSeqTsData.stand}
\alias{phosClusVec}
\alias{protClusVec}
\alias{rnaSeqClusVec}
\title{Multi-omics data sets and clusters.}
\format{An object of class \code{matrix} with 3564 rows and 12 columns.}
\usage{
phosTsData.stand

protTsData.stand

rnaSeqTsData.stand

phosClusVec

protClusVec

rnaSeqClusVec
}
\description{
This dataset was taken from Yang \emph{et al.} 2019 Cell Syst ( 8, 427-445.e10). It is a multiomics data set, comprising transciptomics, proteomics and phosphoproteomics measurements at multiple time points. The clustering of these time series data is also made available. The phosphoproteomics and proteomics data were clustered using Mfuzz, and the transcriptomics data were clustered using STEM (The STEM clustering details can be found in Kaur \emph{et al.} 2020 npj Sys. Bio.).
}
\references{
Yang \emph{et al.} 2019. Available at: \url{http://doi.org/10.1016/j.cels.2019.03.012}
}
\keyword{data}
\keyword{datasets}
