% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderedEvents.r
\name{orderTheEvents}
\alias{orderTheEvents}
\title{Orders clusters based on time of the 50\% percent abundance crossings}
\usage{
orderTheEvents(Tc, clustered, mat_fiftyPoints, test = "wilcox",
  fdrSignif = 0.05)
}
\arguments{
\item{Tc}{A matrix containing time course data.}

\item{clustered}{A fclust object containing the clustering details.}

\item{mat_fiftyPoints}{A matrix containing the event information generated by running the "calc50crossing" function.}

\item{test}{The test to run to determine the orders. The choice is between a parametric or a non-paramtric test:
\describe{
        \item{\code{wilcox}}{Choosing this will result in the computation of 2 NxN matrices containing the 95\% confidence interval value for direction and the p-values, where N is the number of events}
        \item{\code{t-test}}{Choosing this will result in the computation of 2 NxN matrices containing the t-statistic for direction and the p-values}
}}

\item{fdrSignif}{The significant cutoff for FDR correction (performed on a p-value matrix (see above)).}
}
\value{
A figure containing the ordered clusters by occurance of first event is generated. Along with this, returned is a matrix, similar to the mat_fiftyPoints matrix, where an additional column is appended containing the order of occurance of the corresponding event.
}
\description{
Here, the events identified from the centroids are used to generate time-distributions using profiles in the clusters. From these distributions, a matrix (either by running t-test or wilcox) is generated of the p-values and the t-statistic (or the 95\% confidence interval, respectively). The p-values are FDR corrected and are used to define a graph, where events with significant time differences are ordered according to the t-statistic (for t-test) or the 95\% confidence interval (for wilcox test). This matrix is transitively reduced and is used to generate the ordering and the resulting figure.
}
\seealso{
\code{\link[e1071]{cmeans}} for clustering time profiles, \code{\link{calc50crossing}} for identifying events.
}
