% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusters.r
\name{plotClusters_fifty}
\alias{plotClusters_fifty}
\title{Prints a plot of clusters using a monochromatic low opacity gray50 colour.}
\usage{
plotClusters_fifty(Tc, clustered, mat_fiftyPoints, plotNumCol = 5)
}
\arguments{
\item{Tc}{A matrix containing time course data.}

\item{clustered}{A fclust object containing the clustering details}

\item{mat_fiftyPoints}{A matrix containing the event information generated by running the "calc50crossing" function.}

\item{plotNumCol}{Number of columns in the multi-cluster plot}
}
\value{
None
}
\description{
Prints a plot of clusters using a monochromatic low opacity gray50 colour.
}
\seealso{
\code{\link{plotClusters}}
}
