% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterChng.r
\name{plotZP}
\alias{plotZP}
\title{Plot a Z-score heatmap}
\usage{
plotZP(list_concTpSummary, significanceTh = 0.05)
}
\arguments{
\item{list_concTpSummary}{Z-scores and p-values extracted from tukey-constrasts of consecutive time intervals for each cluster. Obtained by running the \code{summaryGetZP} function.}

\item{significanceTh}{Tukey's p-value cutoff to be applied for significance.}
}
\value{
Displays a plot and returns a matrix, with z-scores masked (NA) according to the signficance (p-value) threshold specified.
}
\description{
Plot a heatmap of z-scores, masking the z-scores at insignificant p-values. Rows in the heatmap correspond to clusters, and columns in the heatmap correspond to time intervals.
}
\seealso{
\code{\link{summaryGetZP}}
}
