% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fifty.r
\name{plotZP_fifty}
\alias{plotZP_fifty}
\title{Overlays the 50\% abundance crossings on the Z-score heatmap.}
\usage{
plotZP_fifty(list_concTpSummary, mat_fiftyPoints, significanceTh = 0.5)
}
\arguments{
\item{list_concTpSummary}{Z-scores and p-values extracted from tukey-constrasts for glm's of each cluster. Obtained by running the "summaryGetZP" function.}

\item{mat_fiftyPoints}{A matrix containing the event information generated by running the "calc50crossing" function.}

\item{significanceTh}{P-value cutoff. This value ranges between 0 and 1. Generally, the threshold of p<0.5 is considered significant. This threshold can be reduced to, for example 0.01, 0.05 or 0.001, for only plotting the highly significant z-scores. To plot all z-scores, the treshold can be set to 1.}
}
\value{
Displays a plot and returns a matrix, with z-scores masked (NA) according to the signficance threshold specified.
}
\description{
Overlays the 50\% abundance crossings on the Z-score heatmap.
}
\seealso{
\code{\link{plotZP}}, \code{\link{summaryGetZP}}, \code{\link{calc50crossing}}
}
