% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{plotZP_withEvents}
\alias{plotZP_withEvents}
\title{Plot a Z-score heatmap with events overlaid.}
\usage{
plotZP_withEvents(list_concTpSummary, mat_events, significanceTh = 0.05)
}
\arguments{
\item{list_concTpSummary}{Z-scores and p-values extracted from tukey-constrasts of consecutive time intervals for each cluster. Obtained by running the \code{summaryGetZP} function.}

\item{mat_events}{A matrix containing the event information generated by running the \code{calcEvents} function.}

\item{significanceTh}{P-value cutoff of the Tukey contrasts.}
}
\value{
Displays a plot and returns a matrix, with z-scores masked (NA) according to the signficance threshold specified. The plot also displays events calculated for the centroids.
}
\description{
Plot a heatmap of z-scores, masking the z-scores at insignificant p-values. Rows in the heatmap correspond to clusters, and columns in the heatmap correspond to time intervals. Overlaid on this heatmap are events, where the increasing events are shown in red, and the decreasing events are shown in blue.
}
\seealso{
\code{\link{summaryGetZP}}, \code{\link{calcEvents}}, \code{\link{plotZP}}
}
