% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterChng.r
\name{splitIntoSubMatrices}
\alias{splitIntoSubMatrices}
\title{Split a matrix into cluster based submatrices.}
\usage{
splitIntoSubMatrices(Tc, clusters)
}
\arguments{
\item{Tc}{A matrix containing time course data.}

\item{clusters}{A vector with same length as number of profiles. The cluster labels are assumed to be numbers, starting from 1.}
}
\value{
A list of matrices
}
\description{
Split a matrix into cluster based submatrices.
}
