% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterChng.r
\name{summaryGetZP}
\alias{summaryGetZP}
\title{Summarize clusterChange and get z-scores and p-values.}
\usage{
summaryGetZP(list_resSummaries, totalTimePoints)
}
\arguments{
\item{list_resSummaries}{A list returned by the "calcClusterChng" function.}

\item{totalTimePoints}{Total number of time-points in the experiment.}
}
\value{
A list containing two items, where the first item is a z-score matrix, and the second item is a p-value matrix.
}
\description{
Summarize results obtained by running the "calcClusterChng" function. Specifically this function extracts z-scores and p-values.
}
\seealso{
\code{\link{calcClusterChng}}
}
