% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderedEvents_combined.r
\name{visualizeOrder_combined}
\alias{visualizeOrder_combined}
\title{Visualize the order of events for a multiomics data set.}
\usage{
visualizeOrder_combined(theOrder)
}
\arguments{
\item{theOrder}{A list containing information regarding the order of events and clusters generated via the \code{calculateOrder} function.}
}
\value{
A plot containing the ordered clusters by occurance of first event is generated (see description).
}
\description{
A plot depicting the temporal order of events is generated. The order is shown using two visualizations, event maps (top) and event sparklines (bottom). The event sparkline is a summary of the event map.
}
\seealso{
\code{\link{calculateOrder}}
}
