% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterChng.r
\name{calcClusterChng}
\alias{calcClusterChng}
\title{Order clusters}
\usage{
calcClusterChng(Tc, clustered)
}
\arguments{
\item{Tc}{A matrix containing time course data.}

\item{clustered}{A fclust object containing the clustering details.}
}
\value{
A list (with same length as number of clusters) containing the Tukey summaries.
}
\description{
Create and run generalised linear models and post-hoc tukey contrasts for each cluster.
}
\seealso{
\code{\link[e1071]{cmeans}} for clustering time profiles.
}
