% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{calcEvents}
\alias{calcEvents}
\title{Calculate events for the cluster centroids.}
\usage{
calcEvents(timeRegions, clustered, phosEventTh = 0.5,
  dephosEventTh = 0.5)
}
\arguments{
\item{timeRegions}{A list containing the time regions information.}

\item{clustered}{A fclust object containing the clustering details.}

\item{phosEventTh}{Value between 0 and 1, to define the threshold at which a phosphorylation (or increasing) event occurs. Here, a threshold of 0 corresponds to the minimal value within an interval. An event identified in this way is designated as 1.}

\item{dephosEventTh}{Value between 0 and 1, to define the threshold at which a dephosphorylation (or decreasing) event occurs. A threshold of 0 corresponds to the maximal value within an interval. An event identified in this way is designated as -1.}
}
\value{
A matrix containing the calculated events details for the cluster centroids.
}
\description{
Calculate events for the cluster centroids.
}
\seealso{
\code{\link{getTimeRegionsWithMaximalChange}} to generate \code{timeRegions}, \code{\link[e1071]{cmeans}} to generate \code{clustered}
}
