% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{clusTpDistributions}
\alias{clusTpDistributions}
\title{Densities of standardized abundances at each time for each cluster.}
\usage{
clusTpDistributions(
  Tc,
  clusters,
  outfile = "tpDist.pdf",
  pdfHeight = 40,
  pdfWidth = 50,
  bandwidth = 0.25
)
}
\arguments{
\item{Tc}{A matrix containing time course data.}

\item{clusters}{A vector with same length as number of profiles. The cluster labels are assumed to be numbers, starting from 1.}

\item{outfile}{Name of the output pdf file.}

\item{pdfHeight}{Height of a pdf page.}

\item{pdfWidth}{Width of a pdf page.}

\item{bandwidth}{Bandwidth of the density function.}
}
\description{
A pdf is saved containing the density plots at each time point for each cluster. These distributions should be approximately normally distributed to allow the application of GLMs.
}
\seealso{
\code{\link[e1071]{cmeans}} for clustering time profiles.
}
