% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{missingStats}
\alias{missingStats}
\title{A summary, providing percentage of profiles with missing events in each cluster.}
\usage{
missingStats(Tc, clustered, mat_events, phosEventTh, dephosEventTh)
}
\arguments{
\item{Tc}{A matrix containing time course data.}

\item{clustered}{A fclust object containing the clustering details.}

\item{mat_events}{A matrix containing the event information generated by running the \code{calcEvents} function.}

\item{phosEventTh}{Value between 0 and 1, to define the threshold at which a phosphorylation (or increasing) event occurs. Here, a threshold of 0 corresponds to the minimum value within an interval. (Such an event is designated as 1).}

\item{dephosEventTh}{Value between 0 and 1, to define the threshold at which a dephosphorylation (or decreasing) event occurs. A threshold of 0 corresponds to the maximum value within an interval. (Such an event is designated as -1).}
}
\value{
A matrix providing the number of profiles missing events (1 events per row) for each cluster.
}
\description{
Summary of clusters: providing information regarding the percentage of profiles which do not contain events within the identified time regions.
}
\seealso{
\code{\link{getTimeRegionsWithMaximalChange}}, \code{\link[e1071]{cmeans}}, \code{\link{calcEvents}}
}
