% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusters.r
\name{plotClusters}
\alias{plotClusters}
\title{Plot clusters using a monochromatic low opacity gray colour.}
\usage{
plotClusters(Tc, clustered, plotNumCol = 5)
}
\arguments{
\item{Tc}{A matrix containing time course data.}

\item{clustered}{A fclust object containing the clustering details}

\item{plotNumCol}{Number of cluster-plots to display within a given row.}
}
\value{
None
}
\description{
Using this colour scheme enables the clear visualisation of the density of profiles within each cluster.
}
\seealso{
\code{\link[e1071]{cmeans}} for clustering time profiles and generating the \code{clustered} object.
}
