% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusters.r
\name{plotClusters_withEvents}
\alias{plotClusters_withEvents}
\title{Plots of clusters with events overlaid.}
\usage{
plotClusters_withEvents(Tc, clustered, mat_events, plotNumCol = 5)
}
\arguments{
\item{Tc}{A matrix containing time course data.}

\item{clustered}{A fclust object containing the clustering details}

\item{mat_events}{A matrix containing the event information generated by running the \code{calcEvents} function.}

\item{plotNumCol}{Number of cluster-plots to display within a given row.}
}
\value{
None
}
\description{
Shown overlaid on the cluster plots are events computed for the cluster centroids. The increasing and decreasing events are shown using red and blue, respectively.
}
\seealso{
\code{\link[e1071]{cmeans}} for clustering time profiles and generating the \code{clustered} object, \code{\link{calcEvents}}
}
