% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterChng.r
\name{printZP}
\alias{printZP}
\title{Print changes in consecutive contrasting intervals}
\usage{
printZP(list_concTpSummary)
}
\arguments{
\item{list_concTpSummary}{Z-scores and p-values extracted from tukey-constrasts for glm's of each cluster.}
}
\description{
Prints two matrices, the first containing z-scores, and the second, p-values, where each row corresponds to each cluster, and each column corresponds to the time-interval.
}
\seealso{
\code{\link{summaryGetZP}}
}
