% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderedEvents.r
\name{rearrangeClusters}
\alias{rearrangeClusters}
\title{Rearrange clusters by events}
\usage{
rearrangeClusters(clusters, theOrder)
}
\arguments{
\item{clusters}{A vector with same length as number of profiles. The cluster labels are assumed to be numbers, starting from 1.}

\item{theOrder}{A list containing information regarding the order of events and clusters generated via the \code{calculateOrder} function.}
}
\value{
The returned vector now contains the cluster rearranged according to the computed ordering in \code{theOrder}.
}
\description{
After calculating the order of events, we can now rearrange clusters, so that the first cluster contains the first event and so on.
}
\seealso{
\code{\link[e1071]{cmeans}} for clustering, \code{\link{calculateOrder}} to generate the order.
}
