% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterChng.r
\name{runGlmForEachCluster}
\alias{runGlmForEachCluster}
\title{Run glm for each cluster.}
\usage{
runGlmForEachCluster(list_dfsGlm)
}
\arguments{
\item{list_dfsGlm}{A list of dataframes (for each cluster) of the time series data.}
}
\value{
A list of glm's for each cluster.
}
\description{
Run glm for each cluster.
}
