% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterChng.r
\name{runPostHocTukeyForEachClus}
\alias{runPostHocTukeyForEachClus}
\title{Run tukey post-hoc evaluations for the glm's.}
\usage{
runPostHocTukeyForEachClus(list_lms)
}
\arguments{
\item{list_lms}{A list containing a lm for each cluster.}
}
\value{
A list of tukey evaluations
}
\description{
Run tukey post-hoc evaluations for the glm's.
}
