% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterChng.r, R/calcClusterChng_old.r,
%   R/calcClusterChng_v2.r
\name{splitIntoSubMatrices}
\alias{splitIntoSubMatrices}
\title{Split a matrix into cluster based submatrices.}
\usage{
splitIntoSubMatrices(Tc, clustered)

splitIntoSubMatrices(Tc, clustered)

splitIntoSubMatrices(Tc, clustered)
}
\arguments{
\item{Tc}{A matrix containing time course data.}

\item{clustered}{A fclust object containing the clustering details}

\item{Tc}{A matrix containing time course data.}

\item{clustered}{A fclust object containing the clustering details}

\item{Tc}{A matrix containing time course data.}

\item{clustered}{A fclust object containing the clustering details}
}
\value{
A list of matrices

A list of matrices

A list of matrices
}
\description{
Split a matrix into cluster based submatrices.

Split a matrix into cluster based submatrices.

Split a matrix into cluster based submatrices.
}
