% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterChng.r
\name{summaryGetZP}
\alias{summaryGetZP}
\title{Summarize the Tukey constrasting results for consecutive intervals.}
\usage{
summaryGetZP(list_resSummaries, Tc)
}
\arguments{
\item{list_resSummaries}{A list returned by the \code{calcClusterChng} function.}

\item{Tc}{The time course data}
}
\value{
A list containing two items, where the first item is a z-score matrix, and the second item is a p-value matrix.
}
\description{
Summarize results obtained by running the \code{calcClusterChng} function. Specifically this function extracts z-scores and p-values for consecutive intervals.
}
\seealso{
\code{\link{calcClusterChng}}
}
