% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterChng.r, R/calcClusterChng_old.r,
%   R/calcClusterChng_v2.r
\name{summaryGetZP}
\alias{summaryGetZP}
\title{Summarize clusterChange and get z-scores and p-values.}
\usage{
summaryGetZP(list_resSummaries, Tc)

summaryGetZP(list_resSummaries, Tc)

summaryGetZP(list_resSummaries, Tc)
}
\arguments{
\item{list_resSummaries}{A list returned by the "calcClusterChng" function.}

\item{Tc}{The time course data}

\item{list_resSummaries}{A list returned by the "calcClusterChng" function.}

\item{Tc}{The time course data}

\item{list_resSummaries}{A list returned by the "calcClusterChng" function.}

\item{Tc}{The time course data}
}
\value{
A list containing two items, where the first item is a z-score matrix, and the second item is a p-value matrix.

A list containing two items, where the first item is a z-score matrix, and the second item is a p-value matrix.

A list containing two items, where the first item is a z-score matrix, and the second item is a p-value matrix.
}
\description{
Summarize results obtained by running the "calcClusterChng" function. Specifically this function extracts z-scores and p-values.

Summarize results obtained by running the "calcClusterChng" function. Specifically this function extracts z-scores and p-values.

Summarize results obtained by running the "calcClusterChng" function. Specifically this function extracts z-scores and p-values.
}
\seealso{
\code{\link{calcClusterChng}}

\code{\link{calcClusterChng}}

\code{\link{calcClusterChng}}
}
