% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderedEvents.r, R/orderedEvents_forCombined.r,
%   R/orderedEvents_forFlowChart.r
\name{visualizeOrder}
\alias{visualizeOrder}
\title{Visualize the order of events.}
\usage{
visualizeOrder(theOrder)

visualizeOrder(theOrder)

visualizeOrder(theOrder)
}
\arguments{
\item{theOrder}{A list containing information regarding the order of events and clusters generated via the \code{calculateOrder} function.}
}
\value{
A plot containing the ordered clusters by occurance of first event is generated (see description).

A plot containing the ordered clusters by occurance of first event is generated (see description).

A plot containing the ordered clusters by occurance of first event is generated (see description).
}
\description{
A plot depicting the temporal order of events is generated. The order is shown using two visualizations, event maps (top) and event sparklines (bottom). The event sparkline is a summary of the event map.

A plot depicting the temporal order of events is generated. The order is shown using two visualizations, event maps (top) and event sparklines (bottom). The event sparkline is a summary of the event map.

A plot depicting the temporal order of events is generated. The order is shown using two visualizations, event maps (top) and event sparklines (bottom). The event sparkline is a summary of the event map.
}
\seealso{
\code{\link{calculateOrder}}

\code{\link{calculateOrder}}

\code{\link{calculateOrder}}
}
